Add the MeasuresKokkos class
