+++ pennylane_lightning_kokkos/src/algorithms/CMakeLists.txt
@@ -2,7 +2,7 @@ project(lightning_kokkos_algorithms LANGUAGES CXX)
 set(CMAKE_CXX_STANDARD 17)
 
 
-set(Kokkos_ALGORITHM_FILES AdjointDiffKokkos.hpp AdjointDiffKokkos.cpp CACHE INTERNAL "" FORCE)
+set(Kokkos_ALGORITHM_FILES AdjointDiffKokkos.hpp AdjointDiffKokkos.cpp ExpValFunctors.hpp MeasuresFunctors.hpp MeasuresKokkos.cpp MeasuresKokkos.hpp CACHE INTERNAL "" FORCE)
 add_library(lightning_kokkos_algorithms STATIC ${Kokkos_ALGORITHM_FILES})
 
 target_include_directories(lightning_kokkos_algorithms PUBLIC ${CMAKE_CURRENT_SOURCE_DIR} )
