+++ pennylane_lightning_kokkos/src/algorithms/MeasuresKokkos.cpp
@@ -1,0 +1,5 @@
+#include "MeasuresKokkos.hpp"
+
+// explicit instantiation
+template class Pennylane::Algorithms::MeasuresKokkos<float>;
+template class Pennylane::Algorithms::MeasuresKokkos<double>;
